add xss for security

+++ controller/customFieldController.js
@@ -1,5 +1,8 @@
 const { CustomField } = require("../models/CustomFieldSchema");
 const { validateSuperUser } = require("../middleware/superUserValidator");
+const xss = require("xss");
+const mongoose = require("mongoose");
+
 
 exports.getFields = async (req, res) => {
   const fields = await CustomField.find();
@@ -24,8 +27,14 @@ exports.createField = async (req, res) => {
         .status(401)
         .json({ message: "Options are required for select type" });
     }
+    const fieldData = {
+      label: xss(label),
+      type, // Mongoose Enum will catch if this is invalid
+      required: Boolean(required),
+      options: type === "select" ? options.map((o) => xss(String(o))) : [],
+    };
 
-    let field = await CustomField.create({ label, type, required, options });
+    let field = await CustomField.create(fieldData);
     let userObj = field.toObject();
     return res.status(201).json({
       success: true,
@@ -34,6 +43,9 @@ exports.createField = async (req, res) => {
     });
   } catch (error) {
     console.error("Error:", error);
+    if (error.name === "ValidationError") {
+      return res.status(400).json({ success: false, message: error.message });
+    }
     return res.status(500).json({
       success: false,
       message: "Error while creating field",
@@ -47,6 +59,12 @@ exports.deleteField = async (req, res) => {
     const accessError = validateSuperUser(req, res);
     if (accessError) return;
     const fieldId = req.query.fieldId;
+    if (!mongoose.Types.ObjectId.isValid(fieldId)) {
+      return res.status(400).json({
+        success: false,
+        message: "Invalid ID format. Please provide a valid ObjectId.",
+      });
+    }
     const deletedField = await CustomField.findOneAndDelete({ _id: fieldId });
     if (!deletedField) {
       return res
